% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocard.R
\name{geocard}
\alias{geocard}
\title{Create a geo card}
\usage{
geocard(
  data,
  ref_source,
  geo_level,
  y_domain = NULL,
  y_log_domain = NULL,
  case_fatality_max = 25,
  max_date = NULL,
  min_date = NULL,
  width = NULL,
  height = NULL,
  element_id = NULL
)
}
\description{
Create a geo card
}
