% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogs.R
\name{get_cogs}
\alias{get_cogs}
\title{Compute cognostics for a geographic entity}
\usage{
get_cogs(x, pop, ref_source = NULL)
}
\arguments{
\item{x}{A long-format data frame with columns: "date", "cases", "deaths", "source".}

\item{pop}{Optional population of the geographic entity being plotted (used to populate the "per 100k population" statistics in the card).}

\item{ref_source}{Which source should be used as the reference (default) source? The reference source is the source with which other sources will be compared. It should match one of the values in the "source" column of the provided data.}
}
\description{
Compute cognostics for a geographic entity
}
\examples{
get_cogs(wa_cases, ref_source = "NYT")
}
