% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wa_cases}
\alias{wa_cases}
\title{Example data of COVID-19 case counts from Washington State}
\format{
A data frame with variables:
\describe{
\item{date}{date}
\item{cases}{cumulative number of COVID-19 cases}
\item{deaths}{cumulative number of COVID-19 deaths}
\item{source}{source of statistics (JHU, NYT, FACTS)}
}
}
\usage{
wa_cases
}
\description{
Example data of COVID-19 case counts from Washington State
}
\keyword{datasets}
