% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocard.R
\name{geocard}
\alias{geocard}
\title{Create a geo card}
\usage{
geocard(
  data,
  ref_source,
  geo_level,
  geo_higher_level = NULL,
  y_domain = NULL,
  y_log_domain = NULL,
  case_fatality_max = 25,
  max_date = NULL,
  min_date = NULL,
  width = NULL,
  height = NULL,
  element_id = NULL
)
}
\arguments{
\item{data}{TODO}

\item{ref_source}{TODO}

\item{geo_level}{TODO}

\item{geo_higher_level}{TODO}

\item{y_domain}{TODO}

\item{y_log_domain}{TODO}

\item{case_fatality_max}{TODO}

\item{max_date}{TODO}

\item{min_date}{TODO}

\item{width}{TODO}

\item{height}{TODO}

\item{element_id}{TODO}
}
\description{
Create a geo card
}
