% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogs.R
\name{get_cogs}
\alias{get_cogs}
\title{Compute cognostics for a geographic entity}
\usage{
get_cogs(x, pop)
}
\arguments{
\item{x}{TODO}

\item{pop}{TODO}
}
\description{
Compute cognostics for a geographic entity
}
